Finished M01 Assignment 1
Completed a Jupyter notebook file for "M01 Programming Assignment - Numbers and Types" called assignment_1.ipynb.
Additionally, I added a personal_notes.ipynb which is used for following along with the teacher in class, as well as simply
trying out random coding ideas I have.
Finally, there will be a .vscode folder with data on words to ignore. This is part of the "Code Spell Checker" extension, please ignore